+++ spec/hatchet/pipenv_spec.rb
@@ -5,11 +5,9 @@ require_relative '../spec_helper'
 RSpec.shared_examples 'builds using Pipenv with the requested Python version' do |python_version|
   it "builds with Python #{python_version}" do
     app.deploy do |app|
-      # TODO: Fix the "cp: cannot stat" error here and in the other testcases below (W-7924941).
       expect(clean_output(app.output)).to match(Regexp.new(<<~REGEX))
         remote: -----> Python app detected
         remote: -----> Using Python version specified in Pipfile.lock
-        remote: cp: cannot stat '/tmp/build_.*/requirements.txt': No such file or directory
         remote: -----> Installing python-#{python_version}
         remote: -----> Installing pip 22.2.2, setuptools 63.4.3 and wheel 0.37.1
         remote: -----> Installing dependencies with Pipenv 2020.11.15
@@ -39,18 +37,17 @@ RSpec.describe 'Pipenv support' do
 
     it 'builds with the default Python version using just the Pipfile' do
       app.deploy do |app|
-        expect(clean_output(app.output)).to match(Regexp.new(<<~REGEX))
+        expect(clean_output(app.output)).to include(<<~OUTPUT)
           remote: -----> Python app detected
           remote:  !     No 'Pipfile.lock' found! We recommend you commit this into your repository.
           remote: -----> No Python version was specified. Using the buildpack default: python-#{DEFAULT_PYTHON_VERSION}
           remote:        To use a different version, see: https://devcenter.heroku.com/articles/python-runtimes
-          remote: cp: cannot stat '/tmp/build_.*/requirements.txt': No such file or directory
           remote: -----> Installing python-#{DEFAULT_PYTHON_VERSION}
           remote: -----> Installing pip 22.2.2, setuptools 63.4.3 and wheel 0.37.1
           remote: -----> Installing dependencies with Pipenv 2020.11.15
           remote:        Installing dependencies from Pipfile...
           remote: -----> Installing SQLite3
-        REGEX
+        OUTPUT
       end
     end
   end
@@ -60,17 +57,16 @@ RSpec.describe 'Pipenv support' do
 
     it 'builds with the default Python version' do
       app.deploy do |app|
-        expect(clean_output(app.output)).to match(Regexp.new(<<~REGEX))
+        expect(clean_output(app.output)).to include(<<~OUTPUT)
           remote: -----> Python app detected
           remote: -----> No Python version was specified. Using the buildpack default: python-#{DEFAULT_PYTHON_VERSION}
           remote:        To use a different version, see: https://devcenter.heroku.com/articles/python-runtimes
-          remote: cp: cannot stat '/tmp/build_.*/requirements.txt': No such file or directory
           remote: -----> Installing python-#{DEFAULT_PYTHON_VERSION}
           remote: -----> Installing pip 22.2.2, setuptools 63.4.3 and wheel 0.37.1
           remote: -----> Installing dependencies with Pipenv 2020.11.15
-          remote:        Installing dependencies from Pipfile.lock \\(aad8b1\\)...
+          remote:        Installing dependencies from Pipfile.lock (aad8b1)...
           remote: -----> Installing SQLite3
-        REGEX
+        OUTPUT
       end
     end
   end
@@ -82,18 +78,17 @@ RSpec.describe 'Pipenv support' do
     context 'when using Heroku-18', stacks: %w[heroku-18] do
       it 'builds with the latest Python 2.7' do
         app.deploy do |app|
-          expect(clean_output(app.output)).to match(Regexp.new(<<~REGEX))
+          expect(clean_output(app.output)).to include(<<~OUTPUT)
             remote: -----> Python app detected
             remote: -----> Using Python version specified in Pipfile.lock
             remote:  !     Python 2 has reached its community EOL. Upgrade your Python runtime to maintain a secure application as soon as possible.
             remote:        Learn More: https://devcenter.heroku.com/articles/python-2-7-eol-faq
-            remote: cp: cannot stat '/tmp/build_.*/requirements.txt': No such file or directory
             remote: -----> Installing python-#{LATEST_PYTHON_2_7}
             remote: -----> Installing pip 20.3.4, setuptools 44.1.1 and wheel 0.37.1
             remote: -----> Installing dependencies with Pipenv 2020.11.15
-            remote:        Installing dependencies from Pipfile.lock \\(b8efa9\\)...
+            remote:        Installing dependencies from Pipfile.lock (b8efa9)...
             remote: -----> Installing SQLite3
-          REGEX
+          OUTPUT
         end
       end
     end
@@ -122,16 +117,15 @@ RSpec.describe 'Pipenv support' do
     context 'when using Heroku-18 or Heroku-20', stacks: %w[heroku-18 heroku-20] do
       it 'builds with the latest Python 3.6' do
         app.deploy do |app|
-          expect(clean_output(app.output)).to match(Regexp.new(<<~REGEX))
+          expect(clean_output(app.output)).to include(<<~OUTPUT)
             remote: -----> Python app detected
             remote: -----> Using Python version specified in Pipfile.lock
-            remote: cp: cannot stat '/tmp/build_.*/requirements.txt': No such file or directory
             remote: -----> Installing python-#{LATEST_PYTHON_3_6}
             remote: -----> Installing pip 21.3.1, setuptools 59.6.0 and wheel 0.37.1
             remote: -----> Installing dependencies with Pipenv 2020.11.15
-            remote:        Installing dependencies from Pipfile.lock \\(.*\\)...
+            remote:        Installing dependencies from Pipfile.lock (b8b71b)...
             remote: -----> Installing SQLite3
-          REGEX
+          OUTPUT
         end
       end
     end
@@ -194,18 +188,17 @@ RSpec.describe 'Pipenv support' do
 
     it 'builds with the outdated Python version specified' do
       app.deploy do |app|
-        expect(clean_output(app.output)).to match(Regexp.new(<<~REGEX))
+        expect(clean_output(app.output)).to include(<<~OUTPUT)
           remote: -----> Python app detected
           remote: -----> Using Python version specified in Pipfile.lock
           remote:  !     Python has released a security update! Please consider upgrading to python-#{LATEST_PYTHON_3_10}
           remote:        Learn More: https://devcenter.heroku.com/articles/python-runtimes
-          remote: cp: cannot stat '/tmp/build_.*/requirements.txt': No such file or directory
           remote: -----> Installing python-3.10.4
           remote: -----> Installing pip 22.2.2, setuptools 63.4.3 and wheel 0.37.1
           remote: -----> Installing dependencies with Pipenv 2020.11.15
-          remote:        Installing dependencies from Pipfile.lock \\(e047bc\\)...
+          remote:        Installing dependencies from Pipfile.lock (e047bc)...
           remote: -----> Installing SQLite3
-        REGEX
+        OUTPUT
       end
     end
   end
@@ -246,16 +239,15 @@ RSpec.describe 'Pipenv support' do
 
     it 'builds with the Python version from runtime.txt' do
       app.deploy do |app|
-        expect(clean_output(app.output)).to match(Regexp.new(<<~REGEX))
+        expect(clean_output(app.output)).to include(<<~OUTPUT)
           remote: -----> Python app detected
           remote: -----> Using Python version specified in runtime.txt
-          remote: cp: cannot stat '/tmp/build_.*/requirements.txt': No such file or directory
           remote: -----> Installing python-#{LATEST_PYTHON_3_10}
           remote: -----> Installing pip 22.2.2, setuptools 63.4.3 and wheel 0.37.1
           remote: -----> Installing dependencies with Pipenv 2020.11.15
-          remote:        Installing dependencies from Pipfile.lock \\(75eae0\\)...
+          remote:        Installing dependencies from Pipfile.lock (75eae0)...
           remote: -----> Installing SQLite3
-        REGEX
+        OUTPUT
       end
     end
   end
@@ -287,7 +279,6 @@ RSpec.describe 'Pipenv support' do
           remote: -----> Python app detected
           remote: -----> No Python version was specified. Using the buildpack default: python-#{DEFAULT_PYTHON_VERSION}
           remote:        To use a different version, see: https://devcenter.heroku.com/articles/python-runtimes
-          remote: cp: cannot stat '/tmp/build_.*/requirements.txt': No such file or directory
           remote: -----> Installing python-#{DEFAULT_PYTHON_VERSION}
           remote: -----> Installing pip 22.2.2, setuptools 63.4.3 and wheel 0.37.1
           remote: -----> Installing dependencies with Pipenv 2020.11.15