+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Update pip from 22.1.2 to 22.2.2 for Python 3.7+ ([#1334](https://github.com/heroku/heroku-buildpack-python/pull/1334)).
 - Update setuptools from 60.10.0 to 63.4.3 for Python 3.7+ ([#1334](https://github.com/heroku/heroku-buildpack-python/pull/1334)).
+- Prevent stray `cp: cannot stat ...` error message in build log output when using Pipenv ([#1350](https://github.com/heroku/heroku-buildpack-python/pull/1350)).
 - Remove `BUILD_WITH_GEO_LIBRARIES` sunset messaging ([#1347](https://github.com/heroku/heroku-buildpack-python/pull/1347)).
 - Remove outdated Django version warning ([#1345](https://github.com/heroku/heroku-buildpack-python/pull/1345)).
 - Remove redundant package install warning checks ([#1348](https://github.com/heroku/heroku-buildpack-python/pull/1348)).